[Example] Revert mttkrp & ttmc back to int32 data type

+++ example_new/example_Versal/example_mttkrp/example_project_small/project/aie/adf_kernel.h
@@ -8,9 +8,9 @@
 #define __KERNEL_H__
 using namespace adf;
 
-void kernel_mttkrp0(input_buffer<float, extents<256>>& __restrict in0, input_buffer<float, extents<128>>& __restrict in1, input_buffer<float, extents<256>>& __restrict in2, output_buffer<float, extents<32>>& __restrict out0);
+void kernel_mttkrp0(input_buffer<int32_t, extents<256>>& __restrict in0, input_buffer<int32_t, extents<128>>& __restrict in1, input_buffer<int32_t, extents<256>>& __restrict in2, output_buffer<int32_t, extents<32>>& __restrict out0);
 
-void kernel_mttkrp(input_buffer<float, extents<256>>& __restrict in0, input_buffer<float, extents<128>>& __restrict in1, input_buffer<float, extents<256>>& __restrict in2, input_buffer<float, extents<32>>& __restrict in3, output_buffer<float, extents<32>>& __restrict out0);
+void kernel_mttkrp(input_buffer<int32_t, extents<256>>& __restrict in0, input_buffer<int32_t, extents<128>>& __restrict in1, input_buffer<int32_t, extents<256>>& __restrict in2, input_buffer<int32_t, extents<32>>& __restrict in3, output_buffer<int32_t, extents<32>>& __restrict out0);
 
 
 #endif //__KERNEL_H__/
